requirements.txt
beautifulsoup4==4.8.0
certifi==2019.6.16
chardet==3.0.4
click==7.1.2
colorama==0.4.1
console-menu==0.5.1
docopt==0.6.2
feedparser==5.2.1
Flask==1.1.2
idna==2.8
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
numpy==1.17.0
oauthlib==3.0.2
pandas==0.25.0
praw==7.1.0
prawcore==1.4.0
prettytable==0.7.2
pyfiglet==0.8.post1
PySocks==1.7.0
python-dateutil==2.8.0
python-dotenv==0.13.0
pytz==2019.1
requests==2.22.0
requests-oauthlib==1.2.0
six==1.12.0
soupsieve==1.9.2
stomp.py==6.0.0
tweepy==3.6.0
update-checker==0.17
urllib3==1.25.3
websocket-client==0.57.0
Werkzeug==1.0.1
